+++ tests/BankAccountNormalizer/SixDigitNormalizerTest.php
@@ -7,7 +7,7 @@ use Cs278\BankModulus\BankAccount;
 /**
  * @covers \Cs278\BankModulus\BankAccountNormalizer\SixDigitNormalizer
  */
-final class SixDigitNormalizerTest extends \PHPUnit_Framework_TestCase
+final class SixDigitNormalizerTest extends \PHPUnit\Framework\TestCase
 {
     /** @dataProvider dataSupports */
     public function testSupports($expected, $accountNumber)