+++ tests/Spec/VocaLinkV390Test.php
@@ -4,13 +4,14 @@ namespace Cs278\BankModulus\Spec;
 
 use Cs278\BankModulus\BankAccount;
 use Cs278\BankModulus\BankAccountNormalizer\DefaultNormalizer;
+use Cs278\BankModulus\Exception\CannotValidateException;
 
 /**
  * @covers \Cs278\BankModulus\Spec\VocaLinkV390
  * @covers \Cs278\BankModulus\Spec\VocaLinkV380\Driver
  * @covers \Cs278\BankModulus\Spec\VocaLinkV380\DataV390
  */
-final class VocaLinkV390Test extends \PHPUnit_Framework_TestCase
+final class VocaLinkV390Test extends \PHPUnit\Framework\TestCase
 {
     /**
      * @dataProvider dataCheckValid
@@ -66,10 +67,6 @@ final class VocaLinkV390Test extends \PHPUnit_Framework_TestCase
         }));
     }
 
-    /**
-     * @expectedException \Cs278\BankModulus\Exception\CannotValidateException
-     * @expectedExceptionMessage Unable to determine if the bank details `00-**-00 1******8` are valid or invalid
-     */
     public function testUnknownDetails()
     {
         $checker = new VocaLinkV390();
@@ -77,6 +74,9 @@ final class VocaLinkV390Test extends \PHPUnit_Framework_TestCase
         $bankAccount = $normalizer->normalize(
             new BankAccount('001100', '12345678')
         );
+
+        self::expectException(CannotValidateException::class);
+        self::expectExceptionMessage('Unable to determine if the bank details `00-**-00 1******8` are valid or invalid');
 
         $checker->check($bankAccount);
     }